+++ src/extension.ts
@@ -1,7 +1,7 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
-import {yuqueClone} from './yuque';
+import {yuqueClone, yuqueFetchDocument} from './yuque';
 import { YuqueOutlineProvider } from './yuqueOutline';
 
 // this method is called when your extension is activated
@@ -11,6 +11,8 @@ export function activate(context: vscode.ExtensionContext) {
 	vscode.window.registerTreeDataProvider('yuqueOutline', yuqueOutlineProvider);
 	vscode.commands.registerCommand('yuqueCli.reload', () => yuqueOutlineProvider.load());
 	vscode.commands.registerCommand('yuqueCli.clone', () => yuqueClone());
+	vscode.commands.registerCommand('yuqueCli.fetchDocument',
+		(namespace: string, id: number) => yuqueFetchDocument(namespace, id));
 }
 
 // this method is called when your extension is deactivated